requirements.txt
python3-pip
python-nmap
python3-nmap
colorama
time
